requirements.txt
requests
bs4
openai
django-environ